feat(lint): useAnchorContent for HTML
