+++ internal/compiler/lint/rules/a11y/useAnchorContent.test.md
@@ -2,13 +2,13 @@
 
 **DO NOT MODIFY**. This file has been autogenerated. Run `rome test internal/compiler/lint/rules/harness.test.ts --update-snapshots` to update.
 
-## `jsx-a11y/useAnchorContent`
+## `a11y/useAnchorContent`
 
 ### `0`
 
 ```
 
- lint/jsx-a11y/useAnchorContent/reject/1/file.jsx:1 lint/jsx-a11y/useAnchorContent ━━━━━━━━━━━━━━━━━
+ lint/a11y/useAnchorContent/reject/1/file.jsx:1 lint/a11y/useAnchorContent ━━━━━━━━━━━━━━━━━━━━━━━━━
 
   ✖ Provide screen reader accessible content when using anchor elements.
 
@@ -31,7 +31,7 @@
 
 ```
 
- lint/jsx-a11y/useAnchorContent/reject/2/file.jsx:1 lint/jsx-a11y/useAnchorContent ━━━━━━━━━━━━━━━━━
+ lint/a11y/useAnchorContent/reject/2/file.jsx:1 lint/a11y/useAnchorContent ━━━━━━━━━━━━━━━━━━━━━━━━━
 
   ✖ Provide screen reader accessible content when using anchor elements.
 
@@ -108,5 +108,61 @@
 	<TextWrapper aria-hidden={true} />
 	 visible content
 </a>;
+
+```
+
+### `6`
+
+```
+
+ lint/a11y/useAnchorContent/reject/1/file.html:1 lint/a11y/useAnchorContent ━━━━━━━━━━━━━━━━━━━━━━━━
+
+  ✖ Provide screen reader accessible content when using anchor elements.
+
+    <a></a>
+    ^^^^^^
+
+  ℹ All links on a page should have content that is accessible to screen readers.
+
+
+```
+
+### `6: formatted`
+
+```html
+<a>
+</a>
+
+```
+
+### `7`
+
+```
+
+```
+
+### `7: formatted`
+
+```html
+<a>
+	Anchor Content!
+</a>
+
+```
+
+### `8`
+
+```
+
+```
+
+### `8: formatted`
+
+```html
+<a>
+	<div aria-hidden="true">
+	</div>
+	visible content
+</a>
 
 ```
