+++ internal/compiler/lint/rules/a11y/useAnchorContent.test.rjson
@@ -1,0 +1,25 @@
+[
+	{
+		filename: "file.jsx"
+		invalid: [
+			"<a />"
+			"<a><TextWrapper aria-hidden /></a>"
+	]
+		valid: [
+			"<a>Anchor Content!</a>"
+			"<a><TextWrapper /></a>"
+			"<a dangerouslySetInnerHTML={{ __html: 'foo' }} />"
+			"<a><TextWrapper aria-hidden={true} /> visible content</a>"
+	]
+}
+	{
+		filename: "file.html"
+		invalid: [
+			"<a></a>"
+	]
+		valid: [
+			"<a>Anchor Content!</a>"
+			"<a><div aria-hidden=\"true\"></div>visible content</a>"
+	]
+}
+]
