--- internal/compiler/lint/rules/jsx-a11y/useAnchorContent.test.rjson
@@ -1,11 +1,0 @@
-filename: "file.jsx"
-invalid: [
-	"<a />"
-	"<a><TextWrapper aria-hidden /></a>"
-]
-valid: [
-	"<a>Anchor Content!</a>"
-	"<a><TextWrapper /></a>"
-	"<a dangerouslySetInnerHTML={{ __html: 'foo' }} />"
-	"<a><TextWrapper aria-hidden={true} /> visible content</a>"
-]
